+++ src/main/java/roomescape/reservation/dao/ReservationDao.java
@@ -1,10 +1,10 @@
-package roomescape.reservation.repository;
+package roomescape.reservation.dao;
 
 import roomescape.reservation.domain.Reservation;
 
 import java.util.List;
 
-public interface ReservationRepository {
+public interface ReservationDao {
     List<Reservation> findAll();
 
     Reservation save(Reservation reservation);
