+++ src/main/java/roomescape/reservation/service/ReservationService.java
@@ -2,7 +2,7 @@ package roomescape.reservation.service;
 
 import org.springframework.stereotype.Service;
 import roomescape.reservation.domain.Reservation;
-import roomescape.reservation.repository.ReservationRepository;
+import roomescape.reservation.dao.ReservationRepository;
 
 import java.util.List;
